[BEAM-386] Move UnboundedReadFromBoundedSource to core-construction-java
Now that it's in the construction package, we can delete the
Dataflow-specific clone that we couldn't use from runners-core.
