+++ runners/core-java/pom.xml
@@ -105,11 +105,6 @@
       <artifactId>joda-time</artifactId>
     </dependency>
 
-    <dependency>
-      <groupId>org.slf4j</groupId>
-      <artifactId>slf4j-api</artifactId>
-    </dependency>
-
     <!-- test dependencies -->
 
     <!-- Utilities such as WindowMatchers -->
